Fix int and uint comparison bug

+++ lane_planner/nodes/lane_select/lane_select_core.cpp
@@ -650,7 +650,7 @@ void LaneSelectNode::callbackFromLaneArray(const autoware_msgs::LaneArrayConstPt
     // The number of lanes and waypoints should not change, but to avoid exceptions, the for loop is used to assign to
     // the one with the smaller number.
     uint32_t lanes_size = std::min(tuple_vec_.size(), msg->lanes.size());
-    if (current_lane_idx_ > lanes_size)
+    if (current_lane_idx_ > (int)lanes_size)
     {
       lane_array_id_ = -1;
       current_lane_idx_ = -1;
